requirements.txt
django==3.2.1
Pillow==8.1.0
django-friendship==1.9.1
whitenoise==5.3.0
python-decouple==3.5.0
django-email-verification==0.1.0